feat: Add getCurrentTime function for displaying last update
This commit adds a new file `time.ts` which contains the `getCurrentTime` function. This function returns the current date and time in the format "YYYY-MM-DD HH:mm:ss" using the Asia/Hong_Kong timezone. The function is used in `index.ts` to replace `{last_update}` in the README_TEMPLATE file, allowing for dynamic display of the last update time in the generated README.md file.

Closes #123